demo03/auth-app/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
mediapipe==0.10.0
numpy==1.24.3
pillow==10.0.0